# internal/infrastructure/repository/queries/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package queries

import (
	"time"

	"github.com/google/uuid"
	date "github.com/hardfinhq/go-date"
)

type Group struct {
	ID        uuid.UUID
	Name      string
	DeletedAt *time.Time
}

type Song struct {
	ID          uuid.UUID
	Name        string
	GroupID     uuid.UUID
	ReleaseDate date.Date
	Text        string
	Link        string
	DeletedAt   *time.Time
}